+++ src/ast/passes/semantic_analyser.h
@@ -6,6 +6,10 @@
 #include "collect_nodes.h"
 #include "types.h"
 
+namespace libbpf {
+#include "libbpf/bpf.h"
+} // namespace libbpf
+
 namespace bpftrace::ast {
 
 struct variable {
@@ -74,6 +78,7 @@ public:
   void visit(Sizeof &szof);
   void visit(Offsetof &offof);
   void visit(Map &map);
+  void visit(MapDeclStatement &decl);
   void visit(Variable &var);
   void visit(Binop &binop);
   void visit(Unop &unop);
@@ -180,6 +185,7 @@ private:
   std::map<Node *, CollectNodes<Variable>> for_vars_referenced_;
   std::map<std::string, SizedType> map_val_;
   std::map<std::string, SizedType> map_key_;
+  std::map<std::string, libbpf::bpf_map_type> bpf_map_type_;
 
   uint32_t loop_depth_ = 0;
   bool has_begin_probe_ = false;